+++ homeassistant/components/homekit_controller/__init__.py
@@ -31,7 +31,7 @@ from homeassistant.helpers.typing import ConfigType
 from .config_flow import normalize_hkid
 from .connection import HKDevice, valid_serial_number
 from .const import ENTITY_MAP, KNOWN_DEVICES, TRIGGERS
-from .storage import EntityMapStorage
+from .storage import async_get_entity_storage
 from .utils import async_get_controller, folded_name
 
 _LOGGER = logging.getLogger(__name__)
@@ -240,16 +240,6 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
         raise ConfigEntryNotReady from ex
 
     return True
-
-
-async def async_get_entity_storage(hass: HomeAssistant) -> EntityMapStorage:
-    """Get entity storage."""
-    if ENTITY_MAP in hass.data:
-        map_storage: EntityMapStorage = hass.data[ENTITY_MAP]
-        return map_storage
-    map_storage = hass.data[ENTITY_MAP] = EntityMapStorage(hass)
-    await map_storage.async_initialize()
-    return map_storage
 
 
 async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool:
